fix: serialize announcement state mutations
